+++ application/models/Stocks.php
@@ -5,6 +5,11 @@ class Stocks extends MY_Model
     // constructor
     function __construct()
     {
-        parent::__construct('stocks');
+        parent::__construct('stocks','Code');
     }
+
+    function getAllStocks(){
+        return $this->all();
+    }
+
 }