+++ application/views/stock/index.php
@@ -1,0 +1,30 @@
+<!doctype html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Document</title>
+</head>
+<body>
+
+
+<table class="table">
+    <tr>
+        <th>Code</th>
+        <th>Name</th>
+        <th>Category</th>
+        <th>Value</th>
+    </tr>
+    {stocks}
+    <tr>
+        <td>{Code}</td>
+        <td>{Name}</td>
+        <td>{Category}</td>
+        <td>{Value}</td>
+    </tr>
+    {/stocks}
+</table>
+
+
+
+</body>
+</html>